Add files via upload
Created 5 base classes with limited functionality. User, Application, Catalog, UserType (Enumeration), and a Tester. 

eUserType is an enumeration that has 3 options. May be turned into an advanced enumeration in the future to allow for permission checking.
User contains required information for an account, including an enumeration for the UserType ( to allow for permissions in the future). 
Application contains required information for an application, but does not currently contain a discussion board. 
Catalog simply consists of applications, and a method to display all applications. 
The Tester contains the main method, and currently just serves as a small test of the display of two dummy applications.